+++ src/com/googlecode/propidle/migrations/003-create_highest_revision_table.sql
@@ -1,3 +1,3 @@
 create table highest_revision (
-    revision_number        NUMERIC
+    revision_number        INTEGER
 );
